services/api/pyproject.toml
[project]
name = "chronos-api"
version = "0.1.0"
description = "Chronos middleware/API service"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "pydantic[email]>=2.9",
  "pydantic-settings>=2.6",
  "sqlalchemy>=2.0.36",
  "psycopg[binary]>=3.2",
  "alembic>=1.14",
  "argon2-cffi>=23.1",
  "itsdangerous>=2.2",
  "python-multipart>=0.0.12",
  "PyYAML>=6.0",
  "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.24",
  "httpx>=0.27",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
filterwarnings = [
  "error",
  "ignore::DeprecationWarning",
]
